backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.24
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
